大致完成了写信功能

+++ src/view/new_email.py
@@ -11,23 +11,30 @@ from view import window, check_email
 
 class NewEmail(object):
 
-    def __init__(self, server, sql):
+    def __init__(self, server, database, item_text=None):
         self.__window = window.Window('New Email', 500, 450)
-        # self.__menubar = self.__window.get_menubar()
         self.__root_layout = self.__window.get_root_layout()
-        self.__sql = sql
+        self.__database = database
         self.__server = server
-        self.__sender = server.get_user()
         self.__receiver = StringVar()
         self.__subject = StringVar()
         self.__context = Text()
-        self.__create()
-        self.__start()
+        if item_text is not None:
+            self.__sender = item_text[1]
+            self.__receiver.set(item_text[2])
+            self.__subject.set(item_text[3])
+            self.__create(item_text[4])
+        else:
+            self.__sender = server.get_user()
+            self.__create()
 
-    def __create(self):
+    def __create(self, text=None):
         self.__create_button()
         self.__create_header()
-        self.__create_context()
+        if text is not None:
+            self.__create_context(text)
+        else:
+            self.__create_context()
 
     def __create_button(self):
         frm_button = Frame(self.__root_layout)
@@ -47,22 +54,20 @@ class NewEmail(object):
 
     def __create_header(self):
         frm_header = Frame(self.__root_layout)
-        frm_h_left = Frame(frm_header)
-        Label(frm_h_left, text='From : ', font=('Arial', 12)).grid(row=0, sticky=E)
-        Label(frm_h_left, text='To : ', font=('Arial', 12)).grid(row=1, sticky=E)
-        Label(frm_h_left, text='Subject : ', font=('Arial', 12)).grid(row=2, sticky=W)
-        frm_h_left.pack(side=LEFT)
-        frm_h_right = Frame(frm_header)
-        Label(frm_h_right, text=self.__sender, width=35, font=('Verdana', 12)).grid(row=0)
-        Entry(frm_h_right, textvariable=self.__receiver, width=35, font=('Verdana', 12)).grid(row=1)
-        Entry(frm_h_right, textvariable=self.__subject, width=35, font=('Verdana', 12)).grid(row=2)
-        frm_h_right.pack(side=RIGHT)
+        Label(frm_header, text='From : ', font=('Arial', 12)).grid(row=0, column=0, sticky=E)
+        Label(frm_header, text='To : ', font=('Arial', 12)).grid(row=1, column=0, sticky=E)
+        Label(frm_header, text='Subject : ', font=('Arial', 12)).grid(row=2, column=0, sticky=W)
+        Label(frm_header, text=self.__sender, width=35, font=('Verdana', 12)).grid(row=0, column=1)
+        Entry(frm_header, textvariable=self.__receiver, width=35, font=('Verdana', 12)).grid(row=1, column=1)
+        Entry(frm_header, textvariable=self.__subject, width=35, font=('Verdana', 12)).grid(row=2, column=1)
         frm_header.pack(side=TOP)
 
-    def __create_context(self):
+    def __create_context(self, text=None):
         frm_context = Frame(self.__root_layout)
         Label(frm_context, text='Context', font=('Arial', 12)).pack()
         self.__context = Text(frm_context, width=35, height=12, font=('Verdana', 14))
+        if text is not None:
+            self.__context.insert(0.0, text)
         self.__context.pack()
         frm_context.pack()
 
@@ -74,24 +79,29 @@ class NewEmail(object):
         ti_text = self.__subject.get()
         con_text = self.__context.get(0.0, END)
         if self.__server.send_email(rec_text, con_text, ti_text):
-            messagebox.showinfo('Tips', 'Send email successfully')
+            messagebox.showinfo('Info', 'Send email successfully')
         else:
             messagebox.showwarning('Warning', 'Fail to send email')
 
     def __reset(self):
-        self.__receiver.set('')
-        self.__subject.set('')
-        self.__context.delete(0.0, END)
+        if messagebox.askokcancel('Notice', 'Are you sure to reset all text?'):
+            self.__receiver.set('')
+            self.__subject.set('')
+            self.__context.delete(0.0, END)
 
     def __save(self):
-        sql = 'insert into draft(sender, receiver, subject, context, data)values(%s, %s, %s, %s, %s)'
+        dbsql = 'insert into draft(sender, receiver, subject, context, p_date)values(%s, %s, %s, %s, %s)'
         values = (self.__sender, self.__receiver.get(), self.__subject.get(),
                   self.__context.get(0.0, END), datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
-        self.__sql.exec(sql, values)
+        self.__database.ins_data(dbsql, values)
 
     def __check(self):
-        check_window = check_email.CheckEmail(self.__sql)
+        self.__window.get_window().destroy()
+        global check_window
+        check_window = check_email.CheckEmail(self.__server, self.__database, 'draft')
+        check_window.create()
+        check_window.start()
 
-    def __start(self):
+    def start(self):
         self.__window.get_window().resizable(0, 0)
         self.__window.start()
